Moved mock updates into its own file
