+++ src/tests/clipperUI/panels/changeLogPanel_tests.tsx
@@ -5,6 +5,7 @@ import {ChangeLogPanel} from "../../../scripts/clipperUI/panels/changeLogPanel";
 
 import {HelperFunctions} from "../../helperFunctions";
 import {MithrilUtils} from "../../mithrilUtils";
+import {MockUpdates} from "../../mockUpdates";
 import {TestModule} from "../../testModule";
 
 export class ChangeLogPanelTests extends TestModule {
@@ -14,7 +15,7 @@ export class ChangeLogPanelTests extends TestModule {
 
 	protected tests() {
 		test("For a single update containing multiple changes, those changes should be displayed", () => {
-			let updates = HelperFunctions.getMockUpdates();
+			let updates = MockUpdates.getMockUpdates();
 			let controllerInstance = MithrilUtils.mountToFixture(
 				<ChangeLogPanel updates={updates} />);
 
@@ -31,7 +32,7 @@ export class ChangeLogPanelTests extends TestModule {
 		});
 
 		test("For multiple updates containing multiple changes each, all changes should be displayed", () => {
-			let updates = HelperFunctions.getMockMultipleUpdates();
+			let updates = MockUpdates.getMockMultipleUpdates();
 			let controllerInstance = MithrilUtils.mountToFixture(
 				<ChangeLogPanel updates={updates} />);
 
@@ -48,7 +49,7 @@ export class ChangeLogPanelTests extends TestModule {
 		});
 
 		test("For a single update containing multiple changes, check that their titles and descriptions are being displayed in order", () => {
-			let updates = HelperFunctions.getMockUpdates();
+			let updates = MockUpdates.getMockUpdates();
 			let controllerInstance = MithrilUtils.mountToFixture(
 				<ChangeLogPanel updates={updates} />);
 
@@ -75,7 +76,7 @@ export class ChangeLogPanelTests extends TestModule {
 		});
 
 		test("For an update containing some changes with image urls, check that they get rendered if they have an image, and not rendered if they don't", () => {
-			let updates = HelperFunctions.getMockUpdatesWithSomeImages();
+			let updates = MockUpdates.getMockUpdatesWithSomeImages();
 			let controllerInstance = MithrilUtils.mountToFixture(
 				<ChangeLogPanel updates={updates} />);
 
@@ -100,7 +101,7 @@ export class ChangeLogPanelTests extends TestModule {
 		});
 
 		test("For multiple updates containing multiple changes each, check that their titles and descriptions are being displayed in order", () => {
-			let updates = HelperFunctions.getMockMultipleUpdates();
+			let updates = MockUpdates.getMockMultipleUpdates();
 			let controllerInstance = MithrilUtils.mountToFixture(
 				<ChangeLogPanel updates={updates} />);
